% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFractions.R
\name{clusterFractions}
\alias{clusterFractions}
\title{ClusterFractions}
\usage{
clusterFractions(object, numerator, denominator)
}
\arguments{
\item{object}{A Seurat object with a population to be characterized}

\item{numerator}{}

\item{denominator}{}
}
\description{
Using a Seurat object, calculate the fraction of a population
(e.g. a disease class or treatment group) that can be characterized by
some factor (e.g. cell types)
}
